Now can get_joined and have added very basic tournament metadata scraper
