+++ tennis_new/fetch/tennis_explorer/bin/get_joined.py
@@ -1,0 +1,5 @@
+from tennis_new.fetch.tennis_explorer.combiner import get_joined_from_dailies
+
+
+if __name__ == '__main__':
+    get_joined_from_dailies()
