+++ tennis_new/fetch/tennis_explorer/matches/scraper.py
@@ -101,30 +101,3 @@ class TennisExplorerParser(object):
 
     def write_df(self):
         self.to_df().to_csv(self.path(), index=False)
-
-
-class TennisExplorerTourneyParser(object):
-
-    prefix = "https://www.tennisexplorer.com/"
-    path = Path.joinpath(
-        DATA_PATH,
-        'tourney_summary.csv',
-    )
-
-    def __init__(self, url):
-        self.url = url
-        url_to_try = ('%s%s' % (self.prefix, self.url))
-        self.tree = html.fromstring(requests.get(url_to_try).content)
-        details_elems = self.tree.xpath(".//div[@class='box boxBasic lGray']")
-        assert len(details_elems) == 3
-        details_elem = details_elems[1]
-        self.details = helpers._text(details_elem)  # TODO: Rename _text to public method
-
-    def to_df(self):
-        return pd.DataFrame({
-            'tourney_url': [self.url],
-            'details': [self.details]
-        })
-
-    def write_self(self):
-        self.to_df().to_csv(self.path, mode='a', header=False, index=False)
